requirements.txt
boto3==1.7
python-dotenv==0.8.2
